I added so many things without ever committing them that now I honestly have no idea what I did except for maybe the last thing or two. Added the map feature for viewing events, also added the popup marker on the map. Also started implementing the bottomSheet for search parameters input.
Yes this is not the right branch for all of this, I'll take care of it tomorrow.

+++ lib/widgets/event_card_text_only.dart
@@ -10,7 +10,7 @@ import '../screens/event_details_screen.dart';
 class EventItem extends StatelessWidget {
   final Event event;
   final int index;
-  final double totAmount;
+  final int totAmount;
 
   const EventItem(this.event, this.index, this.totAmount);
